RakutenRoom/requirements.txt
selenium
requests
openai
python-dotenv
PyQt5
